Remove flowrex binary, add --peer flag to flowrcli
- Delete flowr/src/bin/flowrex/ entirely
- Add --peer flag to flowrcli that runs a peer coordinator accepting
  delegated sub-flows from other coordinators on the network
- Keep --server for the existing CLI client-server mode (unchanged)
- Rename tests/flowrex.rs to tests/delegation.rs, keep only the
  end-to-end delegation test using flowrcli --peer as the peer
- Remove all flowrex references from Cargo.toml, README, docs, CI,
  examples, utilities
- Remove flowrex parameter from run_example/test_example utilities
- Update book/running/distributed.md with flowrcli --peer examples
- Update book/introduction/structure.md, flowcore/README.md,
  INSTALLING.md, release workflow
