+++ .github/workflows/release.yml
@@ -128,7 +128,7 @@ jobs:
           fi
 
           FOUND=0
-          for bin in flowc flowrcli flowrgui flowrex flowrdb flowedit; do
+          for bin in flowc flowrcli flowrgui flowrdb flowedit; do
             if [ -f "${RELEASE_DIR}/${bin}${EXT}" ]; then
               cp "${RELEASE_DIR}/${bin}${EXT}" staging/
               echo "Found: ${bin}${EXT}"
