+++ flowr/examples/sequence-of-sequences/main.rs
@@ -1,6 +1,6 @@
 //! A runner for the example using flowrcli
 fn main() {
-    utilities::run_example(file!(), "flowrcli", false, true);
+    utilities::run_example(file!(), "flowrcli", true);
 }
 
 #[cfg(test)]
